Make use of `time.time()` instead of `time.monotonic()`
The seemingly best way of recording the duration of a test run's setup,
actual running time, and teardown, is through the
`pytest_runtest_logreport` hook, which exposes the time at which each
test run hphase started and stopped.

Looking through pytest's source, it seems that these timestamps are
generated using the `time.time()`, as opposed to `time.monotonic()`,
which conflicts with the monotonic timestamping system which we are
currently using.

In order to prevent footguns, use `time.time()`, just as pytest does.